backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pydantic
uvicorn
python-multipart
boto3
aws_lambda_typing
uuid